Move matplotlib drawer into its own file + add test coverage
