+++ projectq/backends/_circuits/__init__.py
@@ -16,5 +16,5 @@ from ._to_latex import to_latex
 from ._plot import to_draw
 
 from ._drawer import CircuitDrawer
-from ._drawer import CircuitDrawerMatplotlib
+from ._drawer_matplotlib import CircuitDrawerMatplotlib
